Made hello_world_2 hello_world, removed unnecessary wrapper and style sheet.
