+++ examples/hello_world/ui/main.html
@@ -1,16 +1,84 @@
 <!DOCTYPE html>
 <html lang="en">
-<head>
+  <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1">
 
-    <link rel="stylesheet" href="style.css">
-</head>
-<body>
-  <h1>Hello World</h1>
-<script>
-    call_py1();
-    call_py2();
-</script>
-</body>
+    <script>
+      function color_text(color) {
+	  document.getElementById("colorable_text").style.color = color;
+      }
+    </script>
+  </head>
+  <body style="background-color:white;">
+    <h1>Hello World</h1>
+    <p id="colorable_text" onclick="call_py('red');">
+      Click me to call a Python function that calls a JavaScript
+      function that colors this paragraph red.
+      Click me to call a Python function that calls a JavaScript
+      function that colors this paragraph red.
+      Click me to call a Python function that calls a JavaScript
+      function that colors this paragraph red.
+      Click me to call a Python function that calls a JavaScript
+      function that colors this paragraph red.
+      Click me to call a Python function that calls a JavaScript
+      function that colors this paragraph red.
+      Click me to call a Python function that calls a JavaScript
+      function that colors this paragraph red.
+      Click me to call a Python function that calls a JavaScript
+      function that colors this paragraph red.
+      Click me to call a Python function that calls a JavaScript
+      function that colors this paragraph red.
+      Click me to call a Python function that calls a JavaScript
+      function that colors this paragraph red.
+      Click me to call a Python function that calls a JavaScript
+      function that colors this paragraph red.
+      Click me to call a Python function that calls a JavaScript
+      function that colors this paragraph red.
+      Click me to call a Python function that calls a JavaScript
+      function that colors this paragraph red.
+      Click me to call a Python function that calls a JavaScript
+      function that colors this paragraph red.
+      Click me to call a Python function that calls a JavaScript
+      function that colors this paragraph red.
+      Click me to call a Python function that calls a JavaScript
+      function that colors this paragraph red.
+      Click me to call a Python function that calls a JavaScript
+      function that colors this paragraph red.
+      Click me to call a Python function that calls a JavaScript
+      function that colors this paragraph red.
+      Click me to call a Python function that calls a JavaScript
+      function that colors this paragraph red.
+      Click me to call a Python function that calls a JavaScript
+      function that colors this paragraph red.
+      Click me to call a Python function that calls a JavaScript
+      function that colors this paragraph red.
+      Click me to call a Python function that calls a JavaScript
+      function that colors this paragraph red.
+      Click me to call a Python function that calls a JavaScript
+      function that colors this paragraph red.
+      Click me to call a Python function that calls a JavaScript
+      function that colors this paragraph red.
+      Click me to call a Python function that calls a JavaScript
+      function that colors this paragraph red.
+      Click me to call a Python function that calls a JavaScript
+      function that colors this paragraph red.
+      Click me to call a Python function that calls a JavaScript
+      function that colors this paragraph red.
+      Click me to call a Python function that calls a JavaScript
+      function that colors this paragraph red.
+      Click me to call a Python function that calls a JavaScript
+      function that colors this paragraph red.
+      Click me to call a Python function that calls a JavaScript
+      function that colors this paragraph red.
+      Click me to call a Python function that calls a JavaScript
+      function that colors this paragraph red.
+      Click me to call a Python function that calls a JavaScript
+      function that colors this paragraph red.
+      Click me to call a Python function that calls a JavaScript
+      function that colors this paragraph red.
+      Click me to call a Python function that calls a JavaScript
+      function that colors this paragraph red.
+    </p>
+  </body>
 </html>
