--- examples/hello_world_2/ui/main.html
@@ -1,90 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    
-    <link rel="stylesheet" href="style.css">
-
-    <script>
-      function call_python_function() {
-	  call_py("red");
-      }
-
-      function color_text_red(color) {
-	  document.getElementById("colorable_text").style.color = color;
-      }
-    </script>
-  </head>
-  <body style="background-color:white;">
-    <h1>Hello World</h1>
-    <p id="colorable_text" onclick="call_python_function()">
-      Click me to call a Python function that calls a JavaScript
-      function that colors this paragraph red.
-      Click me to call a Python function that calls a JavaScript
-      function that colors this paragraph red.
-      Click me to call a Python function that calls a JavaScript
-      function that colors this paragraph red.
-      Click me to call a Python function that calls a JavaScript
-      function that colors this paragraph red.
-      Click me to call a Python function that calls a JavaScript
-      function that colors this paragraph red.
-      Click me to call a Python function that calls a JavaScript
-      function that colors this paragraph red.
-      Click me to call a Python function that calls a JavaScript
-      function that colors this paragraph red.
-      Click me to call a Python function that calls a JavaScript
-      function that colors this paragraph red.
-      Click me to call a Python function that calls a JavaScript
-      function that colors this paragraph red.
-      Click me to call a Python function that calls a JavaScript
-      function that colors this paragraph red.
-      Click me to call a Python function that calls a JavaScript
-      function that colors this paragraph red.
-      Click me to call a Python function that calls a JavaScript
-      function that colors this paragraph red.
-      Click me to call a Python function that calls a JavaScript
-      function that colors this paragraph red.
-      Click me to call a Python function that calls a JavaScript
-      function that colors this paragraph red.
-      Click me to call a Python function that calls a JavaScript
-      function that colors this paragraph red.
-      Click me to call a Python function that calls a JavaScript
-      function that colors this paragraph red.
-      Click me to call a Python function that calls a JavaScript
-      function that colors this paragraph red.
-      Click me to call a Python function that calls a JavaScript
-      function that colors this paragraph red.
-      Click me to call a Python function that calls a JavaScript
-      function that colors this paragraph red.
-      Click me to call a Python function that calls a JavaScript
-      function that colors this paragraph red.
-      Click me to call a Python function that calls a JavaScript
-      function that colors this paragraph red.
-      Click me to call a Python function that calls a JavaScript
-      function that colors this paragraph red.
-      Click me to call a Python function that calls a JavaScript
-      function that colors this paragraph red.
-      Click me to call a Python function that calls a JavaScript
-      function that colors this paragraph red.
-      Click me to call a Python function that calls a JavaScript
-      function that colors this paragraph red.
-      Click me to call a Python function that calls a JavaScript
-      function that colors this paragraph red.
-      Click me to call a Python function that calls a JavaScript
-      function that colors this paragraph red.
-      Click me to call a Python function that calls a JavaScript
-      function that colors this paragraph red.
-      Click me to call a Python function that calls a JavaScript
-      function that colors this paragraph red.
-      Click me to call a Python function that calls a JavaScript
-      function that colors this paragraph red.
-      Click me to call a Python function that calls a JavaScript
-      function that colors this paragraph red.
-      Click me to call a Python function that calls a JavaScript
-      function that colors this paragraph red.
-      Click me to call a Python function that calls a JavaScript
-      function that colors this paragraph red.
-    </p>
-  </body>
-</html>
